requirements.txt
aiohttp
opencv-python
pyserial
numpy
